#7322 refactor TableDisplay structure
